third_party/WebKit/LayoutTests/external/wpt/css/geometry/DOMPoint-001-expected.txt

This is a testharness.js-based test.
PASS testConstructor0
PASS testConstructor1
PASS testConstructor2
PASS testConstructor3
PASS testConstructor4
PASS testConstructor5
FAIL testConstructorDictionary0 assert_equals: Expected value for x is 0 expected 0 but got NaN
FAIL testConstructorDictionary1 assert_equals: Expected value for x is 1 expected 1 but got NaN
FAIL testConstructorDictionary2 assert_equals: Expected value for x is 1 expected 1 but got NaN
FAIL testConstructorDictionary3 assert_equals: Expected value for x is 1 expected 1 but got NaN
FAIL testConstructorDictionary4 assert_equals: Expected value for x is 1 expected 1 but got NaN
FAIL testConstructorDictionary5 assert_equals: Expected value for x is 1 expected 1 but got NaN
FAIL testConstructorDictionary2irregular assert_equals: Expected value for x is 1 expected 1 but got NaN
FAIL testConstructorDictionary2undefined assert_equals: Expected value for x is 1 expected 1 but got NaN
FAIL testConstructorDOMPoint assert_equals: Expected value for x is 1 expected 1 but got NaN
PASS testConstructor2undefined
PASS testConstructorUndefined1
FAIL testConstructorUndefined2 assert_equals: Expected value for y is NaN expected NaN but got 0
PASS DOMPointReadOnly constructor with no values
PASS DOMPointReadOnly constructor with 4 values
PASS testAttributesUndefined
PASS testAttributesNaNInfinity
Harness: the test ran to completion.
